refactor(ui): Extract utils
